+++ lib.php
@@ -647,7 +647,7 @@ class format_topcoll extends format_base {
                     'help_component' => 'format_topcoll',
                     'element_type' => 'tccolourpopup',
                     'element_attributes' => array(
-                        array('tabindex' => -1, 'value' => $defaulttgfgcolour)
+                        array('value' => $defaulttgfgcolour)
                     )
                 );
                 $courseformatoptionsedit['toggleforegroundhovercolour'] = array(
@@ -656,7 +656,7 @@ class format_topcoll extends format_base {
                     'help_component' => 'format_topcoll',
                     'element_type' => 'tccolourpopup',
                     'element_attributes' => array(
-                        array('tabindex' => -1, 'value' => $defaulttgfghvrcolour)
+                        array('value' => $defaulttgfghvrcolour)
                     )
                 );
                 $courseformatoptionsedit['togglebackgroundcolour'] = array(
@@ -665,7 +665,7 @@ class format_topcoll extends format_base {
                     'help_component' => 'format_topcoll',
                     'element_type' => 'tccolourpopup',
                     'element_attributes' => array(
-                        array('tabindex' => -1, 'value' => $defaulttgbgcolour)
+                        array('value' => $defaulttgbgcolour)
                     )
                 );
                 $courseformatoptionsedit['togglebackgroundhovercolour'] = array(
@@ -674,7 +674,7 @@ class format_topcoll extends format_base {
                     'help_component' => 'format_topcoll',
                     'element_type' => 'tccolourpopup',
                     'element_attributes' => array(
-                        array('tabindex' => -1, 'value' => $defaulttgbghvrcolour)
+                        array('value' => $defaulttgbghvrcolour)
                     )
                 );
             } else {